requirements.txt
sqlalchemy
alembic
psycopg2-binary
fastapi
uvicorn[standard]
python-multipart
httpx
pillow
appdirs
python-telegram-bot
argklass
usda_fdc
google-api-python-client
google-auth
easyocr
piper-tts